# openapi.yaml
openapi: 3.1.0
info:
  title: Coupled Modelling API
  version: 1.0.0
  summary: REST API for the coupled multiphysics simulation knowledge base.
  description: |
    HTTP API of the coupled_modelling backend: a knowledge base that stores
    Kratos CoSimulation configurations as OWL/RDF in a GraphDB triple store.

    The backend uses a hybrid architecture:

    * **Direct SPARQL path** — reads and simple mutations (adding, replacing and
      deleting property values, creating and deleting instances) are executed as
      stateless SPARQL queries/updates directly against GraphDB.
    * **In-memory semantic path** — structurally complex workflows (Kratos JSON
      import/export, recursive instance copying, class-axiom inference) load the
      ontology into Owlready2, operate on it in memory and synchronise the
      result back to GraphDB.

    Instances created through the API receive collision-resistant UUID
    identifiers (for example `instance_550e8400-e29b-41d4-a716-446655440000`);
    human-readable names are stored separately as `rdfs:label` annotations.
  contact:
    name: coupled_modelling repository
    url: https://github.com/ldrbmrtv/coupled_modelling
servers:
  - url: http://localhost:5000/api/v1.0
    description: Local development server (uvicorn)
...
paths:
...
  /get_class_hierarchy_metadata/:
    get:
      tags: [explorer]
      operationId: getClassHierarchyMetadata
      summary: List all ontology classes with their direct parents
      description: |
        Returns every project-local OWL class together with the list of its
        direct named superclasses. The frontend builds the class tree from the
        parent arrays (classes with an empty `parents` array are roots).
      responses:
        '200':
          description: Alphabetically sorted class list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassHierarchyEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /search/:
    get:
      tags: [explorer]
      operationId: searchEntities
      summary: Search classes and instances by name or label
      description: |
        Case-insensitive substring search over the local names and labels of
        project classes and instances. Each result group is capped at `limit`;
        matches on a name prefix rank before other matches. Instance results
        carry the same summary shape as `get_class_instance_summaries`.
      parameters:
        - name: q
          in: query
          required: true
          description: Search text (matched as a case-insensitive substring).
          schema:
            type: string
          example: mok
        - name: type
          in: query
          required: false
          description: Restrict results to one entity kind.
          schema:
            type: string
            enum: [all, class, instance]
            default: all
        - name: limit
          in: query
          required: false
          description: Maximum results per group (1-100).
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Matching classes and instances.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
        '400':
          description: Missing or empty `q`, unknown `type`, or `limit` out of range.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /get_class_instance_summaries/:
    get:
      tags: [explorer]
      operationId: getClassInstanceSummaries
      summary: List instance summaries for a class
      description: |
        Returns one summary per instance of the given class (including
        instances of its subclasses, via `rdf:type/rdfs:subClassOf*`). Each
        summary carries the resolved preferred label (English first, then
        untagged, then other languages), the direct types, and a compact
        preview of up to three property values for display under the instance
        title. When `class` is omitted, summaries for all instances are
        returned.
      parameters:
        - name: class
          in: query
          required: false
          description: Local class name to list instances for (for example `solvers`). Omit for all instances.
          schema:
            type: string
          example: solvers
      responses:
        '200':
          description: Instance summaries sorted by label.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstanceSummary'
        '400':
          description: The class does not exist in GraphDB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /get_class_metadata/:
    get:
      tags: [explorer]
      operationId: getClassMetadata
      summary: Get class descriptions, hierarchy links and restriction axioms
      description: |
        Returns metadata backing the class inspector pane: `rdfs:comment` /
        `skos:definition` descriptions, direct superclasses and subclasses,
        named equivalent classes, and the OWL restriction axioms asserted on
        the class (existential/universal/value restrictions and unqualified or
        qualified cardinality restrictions, including intersection targets
        expanded into their named member classes).
      parameters:
        - name: class
          in: query
          required: true
          description: Local class name (for example `coupled_system`).
          schema:
            type: string
          example: coupled_system
      responses:
        '200':
          description: Class metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassMetadata'
        '400':
          description: Missing `class` parameter, or the class does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /get_instance_property_metadata/:
    get:
      tags: [explorer]
      operationId: getInstancePropertyMetadata
      summary: Get an instance's direct properties with type metadata
      description: |
        Returns the direct outgoing statements of an instance, grouped per
        property and sorted, with each value tagged as an `object` reference
        (id plus resolved label) or a typed `literal` (JSON-native value plus
        XSD datatype). Property names are returned without the `has_` prefix.
      parameters:
        - name: instance
          in: query
          required: true
          description: Local instance identifier.
          schema:
            type: string
          example: instance_550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Instance metadata for the inspector pane.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceMetadata'
        '400':
          description: Missing `instance` parameter, or the identifier is not an existing individual (classes and properties do not count).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /create_instance/:
    post:
      tags: [mutations]
      operationId: createInstance
      summary: Create an instance linked to a parent (direct SPARQL)
      description: |
        Creates a new UUID-named instance of class `property`, links it to the
        existing `parent` instance via the corresponding `has_<property>`
        object property, and optionally asserts initial property values from
        `data`. String values that match the label of an existing instance of
        the property's class are resolved to object references; creating
        nested individuals inline is rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [property, parent]
              properties:
                property:
                  type: string
                  description: Class name of the new instance (also selects the `has_*` linking property).
                  example: solvers
                parent:
                  type: string
                  description: Identifier of the existing parent instance.
                data:
                  $ref: '#/components/schemas/PropertyDataMap'
      responses:
        '201':
          description: Identifier of the created instance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceId'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /create_class_instance/:
    post:
      tags: [mutations]
      operationId: createClassInstance
      summary: Create a standalone labelled instance (direct SPARQL)
      description: Creates a new UUID-named instance of the given class with an `rdfs:label`, without linking it to a parent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [class, label]
              properties:
                class:
                  type: string
                  description: Class to instantiate.
                  example: solvers
                label:
                  type: string
                  description: Human-readable label stored as `rdfs:label`.
                  example: Airfoil fluid solver
      responses:
        '201':
          description: Identifier of the created instance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceId'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /add_values/:
    post:
      tags: [mutations]
      operationId: addValues
      summary: Add property values to an instance (direct SPARQL)
      description: |
        Inserts the given property values for an existing instance in a single
        SPARQL update transaction. List values produce one triple per element.
        Plain string values are kept as literals for datatype properties; for
        object properties they are resolved to an existing instance of the
        property's class by label, or a new labeled instance is created when no
        match exists. Explicit `instance*` references must already exist.
      requestBody:
        $ref: '#/components/requestBodies/InstanceDataBody'
      responses:
        '201':
          $ref: '#/components/responses/EmptyCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /replace_values/:
    post:
      tags: [mutations]
      operationId: replaceValues
      summary: Replace property values on an instance (direct SPARQL)
      description: |
        Replaces the existing values of each property in `data` with the given
        new values in a single SPARQL update transaction. An empty list (`[]`)
        removes all values of that property.
      requestBody:
        $ref: '#/components/requestBodies/InstanceDataBody'
      responses:
        '201':
          $ref: '#/components/responses/EmptyCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /replace_properties/:
    post:
      tags: [mutations]
      operationId: replaceProperties
      summary: Replace an instance's property set (direct SPARQL)
      description: |
        Removes the instance's existing `has_*` and `rdfs:label` statements and
        asserts the given data instead, preserving structural metadata
        (`rdf:type`, `rdfs:comment`).
      requestBody:
        $ref: '#/components/requestBodies/InstanceDataBody'
      responses:
        '201':
          $ref: '#/components/responses/EmptyCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /replace_value/:
    post:
      tags: [mutations]
      operationId: replaceValue
      summary: Atomically replace one specific property value (direct SPARQL)
      description: >
        Replaces one stored value with another in a single SPARQL update. The old value
        is matched by value equality (dangling references allowed); the new value keeps
        its exact datatype and language tag. When the old value no longer exists, the
        update is a no-op. Replacing an object value removes the old link only; the
        previous target is not garbage-collected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instance, property, old_value, new_value]
              properties:
                instance:
                  type: string
                  description: Instance identifier.
                property:
                  type: string
                  description: Property name (without the `has_` prefix).
                old_value:
                  description: The currently stored value to replace.
                  oneOf:
                    - $ref: '#/components/schemas/ObjectValueTarget'
                    - $ref: '#/components/schemas/LiteralValueTarget'
                new_value:
                  description: The replacement value.
                  oneOf:
                    - $ref: '#/components/schemas/ObjectValueTarget'
                    - $ref: '#/components/schemas/LiteralValueTarget'
      responses:
        '201':
          $ref: '#/components/responses/EmptyCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /delete_value/:
    post:
      tags: [mutations]
      operationId: deleteValue
      summary: Delete a single property value (direct SPARQL)
      description: |
        Removes the triple. When the value is an object link and `cascade` is set
        (the default), the linked instance's owned subtree is deleted as well if
        nothing else reaches it afterwards: the same ownership rule as
        `delete_instance`. A coupled system, a target still linked from elsewhere,
        a link that is not stored, and the instance holding the link are never
        collected. The removal and the collection run in one transactional
        update. Use `get_value_deletion_preview` to see the outcome first.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instance, property, value]
              properties:
                instance:
                  type: string
                  description: Instance identifier.
                property:
                  type: string
                  description: Property name (without the `has_` prefix).
                  example: parallel_type
                value:
                  description: >
                    The value to delete. The typed forms delete the exact triple (correct
                    datatype serialization); a bare scalar is matched by serialization guess.
                  oneOf:
                    - $ref: '#/components/schemas/ObjectValueTarget'
                    - $ref: '#/components/schemas/LiteralValueTarget'
                    - $ref: '#/components/schemas/ScalarValue'
                cascade:
                  type: boolean
                  default: true
                  description: Also delete the unlinked instance's owned subtree when nothing else reaches it.
      responses:
        '200':
          description: The value was deleted.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnlinkResult'
                  - type: object
                    required: [status]
                    properties:
                      status:
                        type: string
                        const: success
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /get_value_deletion_preview/:
    get:
      tags: [explorer]
      operationId: getValueDeletionPreview
      summary: Preview what removing an object link would delete
      description: |
        Read-only counterpart of `delete_value` for an object link: whether the
        linked instance would be deleted with its owned subtree (nothing else
        reaches it afterwards) or kept.
      parameters:
        - name: instance
          in: query
          required: true
          description: Identifier of the instance holding the link.
          schema:
            type: string
          example: instance_550e8400-e29b-41d4-a716-446655440000
        - name: property
          in: query
          required: true
          description: Property name (without the `has_` prefix).
          schema:
            type: string
          example: solver_settings
        - name: target
          in: query
          required: true
          description: Identifier of the linked instance.
          schema:
            type: string
      responses:
        '200':
          description: The unlink preview.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnlinkResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /delete_values/:
    post:
      tags: [mutations]
      operationId: deleteValues
      summary: Delete all values of the given properties (direct SPARQL)
      description: Removes the triples only; unlinked instances are not garbage-collected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instance, properties]
              properties:
                instance:
                  type: string
                  description: Instance identifier.
                properties:
                  type: array
                  description: Property names whose values are removed (without the `has_` prefix).
                  items:
                    type: string
                  example: [echo_level, parallel_type]
      responses:
        '201':
          $ref: '#/components/responses/EmptyCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/GraphDBUnavailable'
        '500':
          $ref: '#/components/responses/UnexpectedError'
...
  /save_onto/:
    post:
      tags: [files]
      operationId: saveOnto
      summary: Persist the in-memory ontology to GraphDB
      description: Pushes the current in-memory Owlready2 model to the GraphDB named graph.
      responses:
        '201':
          $ref: '#/components/responses/EmptyCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
...
  schemas:
...
    ClassHierarchyEntry:
      type: object
      required: [class, parents]
      properties:
        class:
          type: string
          description: Local class name.
          example: convergence_accelerators
        parents:
          type: array
          description: Direct named superclasses (empty for root classes).
          items:
            type: string
...